feat: add multiselect v2 as defualt, cleanup old
