+++ packages/ui/src/components/index.ts
@@ -62,7 +62,6 @@ export * from './switch'
 export * from './meter'
 export * from './filters'
 export * from './more-actions-tooltip'
-export * from './multi-select'
 export * from './split-button'
 export * from './tooltip'
 export * from './navigation-menu'
@@ -80,7 +79,7 @@ export * from './informer'
 export * from './tag'
 export * from './sorts'
 export * from './entity-form-layout'
-export * as MultiSelectV2 from './multi-select-v2'
+export * from './multi-select'
 export * from './card-select'
 
 export * as NodeGroup from './node-group'
